% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.R
\name{cdf}
\alias{cdf}
\title{Cumulative distribution function value, F(x), at each element, x_i, of x.}
\usage{
cdf(x, emm)
}
\description{
Cumulative distribution function value, F(x), at each element, x_i, of x.
}
