% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm.fit.R
\name{emm.fit}
\alias{emm.fit}
\title{Exponential Mixture Model}
\usage{
emm.fit(x, k = 10, n_init = 10, n_iter = 1000, seed = 1)
}
\arguments{
\item{x}{Numerical vector. Data.}

\item{k}{int, optional. Number of components initially given.}

\item{n_init}{int, optional. EM algorithm is run n_init times with different initial conditions. The estimated parameters that maximize the joint likelihood are selected.}

\item{n_iter}{int, optional. Number of iterations in the EM algorithm.}

\item{seed}{int, 1 (default). Seed of random number generator.}
}
\value{
List with pi (vector - Mixing weight of the individual exponential distribution in the estimated EMM. The vector length is k_final.) and mu (vector - Mean parameter of the individual exponential distribution in the estimated EMM. The vector length is k_final.).
}
\description{
Fit an EMM to data 'x'. EM algorithm is used for fitting. The EM algorithm is run several times to avoid local maxima as much as possible.
}
\keyword{emm,}
\keyword{exponential}
\keyword{mixture}
\keyword{model}
